+++ packages/brick-dll/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.34](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@1.3.33...@easyops/brick-dll@1.3.34) (2020-09-03)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [1.3.33](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@1.3.32...@easyops/brick-dll@1.3.33) (2020-09-02)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [1.3.32](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@1.3.31...@easyops/brick-dll@1.3.32) (2020-09-01)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
 # [2.0.0-next.6](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@2.0.0-next.4...@easyops/brick-dll@2.0.0-next.6) (2020-08-28)
 
 ### Bug Fixes