+++ packages/create-next-repo/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.45](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@0.13.44...@easyops/create-next-repo@0.13.45) (2020-09-03)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [0.13.44](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@0.13.43...@easyops/create-next-repo@0.13.44) (2020-09-02)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
+## [0.13.43](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@0.13.42...@easyops/create-next-repo@0.13.43) (2020-09-01)
+
+**Note:** Version bump only for package @easyops/create-next-repo
+
 # [1.0.0-next.8](https://git.easyops.local/anyclouds/next-core/compare/@easyops/create-next-repo@1.0.0-next.6...@easyops/create-next-repo@1.0.0-next.8) (2020-08-28)
 
 ### Bug Fixes