+++ packages/dev-dependencies/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.16](https://git.easyops.local/anyclouds/next-core/compare/@easyops/dev-dependencies@0.7.15...@easyops/dev-dependencies@0.7.16) (2020-09-03)
+
+**Note:** Version bump only for package @easyops/dev-dependencies
+
+## [0.7.15](https://git.easyops.local/anyclouds/next-core/compare/@easyops/dev-dependencies@0.7.14...@easyops/dev-dependencies@0.7.15) (2020-09-02)
+
+**Note:** Version bump only for package @easyops/dev-dependencies
+
+## [0.7.14](https://git.easyops.local/anyclouds/next-core/compare/@easyops/dev-dependencies@0.7.13...@easyops/dev-dependencies@0.7.14) (2020-09-01)
+
+**Note:** Version bump only for package @easyops/dev-dependencies
+
 # [1.0.0-next.8](https://git.easyops.local/anyclouds/next-core/compare/@easyops/dev-dependencies@1.0.0-next.6...@easyops/dev-dependencies@1.0.0-next.8) (2020-08-28)
 
 ### Bug Fixes